fix: Handle ParsleyNode types with Rust equivalent and handles failed slicing

+++ src-tauri/src/debug_tree.rs
@@ -25,13 +25,13 @@ pub struct DebugNode {
     pub name: String, /*The internal (default) or user-defined name of the parser */
     pub internal: String, /*The internal name of the parser */
     pub success: bool, /* Whether the parser was successful */
-    pub child_id: Option<usize>, /* The unique child number of this node */
+    pub child_id: Option<u32>, /* The unique child number of this node */
     pub input: String, /* The input string passed to the parser */
     pub children: Vec<DebugNode>, /* The children of this node */
 }
 
 impl DebugNode {
-    pub fn new(name: String, internal: String, success: bool, child_id: Option<usize>, input: String, children: Vec<DebugNode>) -> Self {
+    pub fn new(name: String, internal: String, success: bool, child_id: Option<u32>, input: String, children: Vec<DebugNode>) -> Self {
         DebugNode {
             name,
             internal,
